+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
+- Best ignore behaviour. Pass the fileMeta to the ignore function. [#33](https://github.com/geut/postcss-copy/issues/33)
 - Change default template with `'[hash].[ext]'`.
 - Replace minimatch by micromatch.
 - Use keepachangelog format.